Flatten the engine::input::input module
